Move column naming and conversion outside read_csv
In hfdata.py, column naming and column conversion is now the same across
different datasource types. In addition to the gained uniformity, this
also simplifies specifying the converter func since the func argument is
now always a number (not a string.  Performance-wise, converting after
reading doesn't seem to make much difference. Am leaving usecols inside
the call to read_csv because it does seem to be faster than reading the
whole dataframe